Created sequences and triggers for status lookup tables
git-svn-id: svn+ssh://bandit.bloomington.in.gov/srv/svn/master_address/trunk@110 c0873d58-3416-0410-9558-ca141da00fa8

+++ scripts/mysql.sql
@@ -77,7 +77,7 @@ create table building_types_master (
 );
 
 create table buildings_status_lookup (
-	status_code tinyint unsigned not null primary key auto_increment,
+	status_code int unsigned not null primary key auto_increment,
 	description varchar2(240) not null
 );
